refactor(petri): spire-backed number types; drop lib.number dep
Replace hydrozoa's PositiveInt/NonNegativeInt so the framework depends only on
spire+cats (no lib.number) — a prerequisite for publishing it standalone. Arc
weights (Inscription.Weighted.coeff, Arc.Syntax.Weighted) get a new framework
type Positive (spire Natural, >= 1): per ISO 15909-1 the flow relation holds only
positive-weight arcs, so a 0 weight is the absence of an arc, not a present arc
moving nothing (inert at firing but a structurally phantom edge that would
mislead incidence/siphon/trap analysis). The genuinely non-negative fields
(Place bound/radius, Transition priority/width/height) become spire Natural.

+++ src/main/scala/hydrozoa/lib/petri/hlpn/Binding.scala
@@ -61,7 +61,7 @@ object Binding:
             Multiset(entries.to(SortedMap))
         term match
             case Inscription.Weighted(coeff, color) =>
-                evalColor(color, b).map(c => bag(List(c -> SafeLong(coeff.toInt))))
+                evalColor(color, b).map(c => bag(List(c -> SafeLong(coeff.toLong))))
             case Inscription.Union(l, r) =>
                 for lb <- evalInscription(l, b); rb <- evalInscription(r, b)
                 yield lb.combineWith(rb)((x, y) => x + y)
